+++ tests/run_examples.py
@@ -82,7 +82,7 @@ def test_demo1():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo1.yaml', logger=logger)[0]
     print('Running demo1.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     # There is no assert at the end of this one, since they are not expected to be identical
     # due to the lack of a specified seed.  This just checks for syntax errors.
 
@@ -98,7 +98,7 @@ def test_demo2():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo2.yaml', logger=logger)[0]
     print('Running demo2.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     assert check_same('output/demo2.fits', 'output_yaml/demo2.fits')
 
 @timer
@@ -113,7 +113,7 @@ def test_demo3():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo3.yaml', logger=logger)[0]
     print('Running demo3.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     assert check_same('output/demo3.fits', 'output_yaml/demo3.fits')
     assert check_same('output/demo3_epsf.fits', 'output_yaml/demo3_epsf.fits')
 
@@ -129,7 +129,7 @@ def test_demo4():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo4.yaml', logger=logger)[0]
     print('Running demo4.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     assert check_same('output/multi.fits', 'output_yaml/multi.fits')
 
 @timer
@@ -144,7 +144,7 @@ def test_demo5():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo5.yaml', logger=logger)[0]
     print('Running demo5.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     assert check_same('output/g08_psf.fits', 'output_yaml/g08_psf.fits')
     assert check_same('output/g08_gal.fits', 'output_yaml/g08_gal.fits')
 
@@ -160,9 +160,9 @@ def test_demo6():
     logger = logging.getLogger('galsim')
     configs = galsim.config.ReadConfig('demo6.yaml', logger=logger)
     print('Running demo6.yaml pass #1')
-    galsim.config.Process(configs[0], logger=logger)
+    galsim.config.Process(configs[0], logger=logger, except_abort=True)
     print('Running demo6.yaml pass #2')
-    galsim.config.Process(configs[1], logger=logger)
+    galsim.config.Process(configs[1], logger=logger, except_abort=True)
     assert check_same('output/psf_real.fits', 'output_yaml/psf_real.fits')
     assert check_same('output/cube_real.fits', 'output_yaml/cube_real.fits')
 
@@ -180,7 +180,7 @@ def test_demo7():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo7.yaml', logger=logger)[0]
     print('Running demo7.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     # gzip class in python 2.6 doesn't implement context correctly.  So do that one manually,
     # even though with gzip.open(...) as f_in would work fine on 2.7+
     f_in = gzip.open('output/cube_phot.fits.gz', 'rb')
@@ -205,9 +205,9 @@ def test_demo8():
     logger = logging.getLogger('galsim')
     configs = galsim.config.ReadConfig('demo8.yaml', logger=logger)
     print('Running demo8.yaml pass #1')
-    galsim.config.Process(configs[0], logger=logger)
+    galsim.config.Process(configs[0], logger=logger, except_abort=True)
     print('Running demo8.yaml pass #2')
-    galsim.config.Process(configs[1], logger=logger)
+    galsim.config.Process(configs[1], logger=logger, except_abort=True)
     assert check_same('output/bpd_single.fits', 'output_yaml/bpd_single.fits')
     assert check_same('output/bpd_multi.fits', 'output_yaml/bpd_multi.fits')
 
@@ -227,11 +227,14 @@ def test_demo9():
     config = galsim.config.ReadConfig('json/demo9.json', logger=logger)[0]
     print('Running demo9.json')
     new_params = { 'output.skip' : { 'type' : 'List', 'items' : [0,0,0,0,0,1] } }
-    galsim.config.Process(config, logger=logger, new_params=new_params, njobs=3, job=1)
-    galsim.config.Process(config, logger=logger, new_params=new_params, njobs=3, job=2)
-    galsim.config.Process(config, logger=logger, new_params=new_params, njobs=3, job=3)
+    galsim.config.Process(config, logger=logger, new_params=new_params, njobs=3, job=1,
+                          except_abort=True)
+    galsim.config.Process(config, logger=logger, new_params=new_params, njobs=3, job=2,
+                          except_abort=True)
+    galsim.config.Process(config, logger=logger, new_params=new_params, njobs=3, job=3,
+                          except_abort=True)
     new_params = { 'output.noclobber' : True }
-    galsim.config.Process(config, logger=logger, new_params=new_params)
+    galsim.config.Process(config, logger=logger, new_params=new_params, except_abort=True)
     for dir_num in range(1,5):
         for file_num in range(5):
             file_name = 'nfw%d/cluster%04d.fits'%(dir_num, file_num)
@@ -251,7 +254,7 @@ def test_demo10():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo10.yaml', logger=logger)[0]
     print('Running demo10.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     assert check_same('output/power_spectrum.fits', 'output_yaml/power_spectrum.fits')
 
 @timer
@@ -266,7 +269,7 @@ def test_demo11():
     logger = logging.getLogger('galsim')
     config = galsim.config.ReadConfig('demo11.yaml', logger=logger)[0]
     print('Running demo11.yaml')
-    galsim.config.Process(config, logger=logger)
+    galsim.config.Process(config, logger=logger, except_abort=True)
     assert check_same('output/tabulated_power_spectrum.fits.fz',
                       'output_yaml/tabulated_power_spectrum.fits.fz')
 
@@ -315,7 +318,7 @@ def test_des():
         configs = galsim.config.ReadConfig('draw_psf.yaml', logger=logger)
         for config in configs:
             config['output']['nfiles'] = 1
-            galsim.config.Process(config, logger=logger)
+            galsim.config.Process(config, logger=logger, except_abort=True)
         assert check_same('output/DECam_00154912_01_psfex_image.fits',
                           'output_yaml/DECam_00154912_01_psfex_image.fits')
         assert check_same('output/DECam_00154912_01_fitpsf_image.fits',
@@ -328,16 +331,16 @@ def test_des():
         config['input']['cosmos_catalog']['file_name'] = '../data/real_galaxy_catalog_23.5_example.fits'
         del config['input']['cosmos_catalog']['sample']
         config['input']['des_wcs']['bad_ccds'] = list(range(2,63))  # All but CCD 1
-        galsim.config.Process(config, logger=logger)
+        galsim.config.Process(config, logger=logger, except_abort=True)
 
         input_cosmos = config['input']['cosmos_catalog'] # Save example COSMOS catalog spec.
         config = galsim.config.ReadConfig('blend.yaml', logger=logger)[0]
-        galsim.config.Process(config, logger=logger)
+        galsim.config.Process(config, logger=logger, except_abort=True)
 
         config = galsim.config.ReadConfig('blendset.yaml', logger=logger)[0]
         config['input']['cosmos_catalog'] = input_cosmos
         config['input']['des_psfex']['file_name']['num'] = 1
-        galsim.config.Process(config, logger=logger)
+        galsim.config.Process(config, logger=logger, except_abort=True)
 
     finally:
         os.chdir(original_dir)
@@ -371,7 +374,8 @@ def test_great3():
             else:
                 new_params = p2
             for config in configs:
-                galsim.config.Process(config, logger=logger, new_params=new_params)
+                galsim.config.Process(config, logger=logger, new_params=new_params,
+                                      except_abort=True)
     finally:
         os.chdir(original_dir)